+++ servicetalk-buffer-internal/build.gradle
@@ -1,0 +1,28 @@
+/*
+ * Copyright © 2021 Apple Inc. and the ServiceTalk project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+apply plugin: "io.servicetalk.servicetalk-gradle-plugin-internal-library"
+
+dependencies {
+    implementation project(":servicetalk-annotations")
+    implementation project(":servicetalk-buffer-api")
+
+    implementation "com.google.code.findbugs:jsr305:$jsr305Version"
+
+    testImplementation "junit:junit:$junitVersion"
+    testImplementation project(":servicetalk-buffer-netty")
+    testFixturesImplementation "org.hamcrest:hamcrest-library:$hamcrestVersion"
+}
