+++ servicetalk-buffer-internal/src/testFixtures/java/io/servicetalk/buffer/internal/Matchers.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2020 Apple Inc. and the ServiceTalk project authors
+ * Copyright © 2020-2021 Apple Inc. and the ServiceTalk project authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.servicetalk.http.api;
+package io.servicetalk.buffer.internal;
 
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 
-import static io.servicetalk.http.api.CharSequences.contentEquals;
+import static io.servicetalk.buffer.internal.CharSequences.contentEquals;
 import static java.util.Objects.requireNonNull;
 
 /**
