+++ servicetalk-http-api/src/main/java/io/servicetalk/http/api/DefaultHttpHeaders.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018 Apple Inc. and the ServiceTalk project authors
+ * Copyright © 2018, 2021 Apple Inc. and the ServiceTalk project authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
  */
 package io.servicetalk.http.api;
 
+import io.servicetalk.buffer.internal.CharSequences;
+
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -23,9 +25,9 @@ import java.util.NoSuchElementException;
 import java.util.Set;
 import javax.annotation.Nullable;
 
-import static io.servicetalk.http.api.CharSequences.caseInsensitiveHashCode;
-import static io.servicetalk.http.api.CharSequences.contentEquals;
-import static io.servicetalk.http.api.CharSequences.contentEqualsIgnoreCase;
+import static io.servicetalk.buffer.internal.CharSequences.caseInsensitiveHashCode;
+import static io.servicetalk.buffer.internal.CharSequences.contentEquals;
+import static io.servicetalk.buffer.internal.CharSequences.contentEqualsIgnoreCase;
 import static io.servicetalk.http.api.DefaultHttpSetCookie.parseSetCookie;
 import static io.servicetalk.http.api.HeaderUtils.DEFAULT_HEADER_FILTER;
 import static io.servicetalk.http.api.HeaderUtils.domainMatches;
