+++ servicetalk-http-api/src/main/java/io/servicetalk/http/api/DefaultHttpSetCookie.java
@@ -15,18 +15,20 @@
  */
 package io.servicetalk.http.api;
 
+import io.servicetalk.buffer.internal.AsciiBuffer;
+
 import java.util.Objects;
 import javax.annotation.Nullable;
 
-import static io.servicetalk.http.api.CharSequences.caseInsensitiveHashCode;
-import static io.servicetalk.http.api.CharSequences.contentEqualsIgnoreCase;
-import static io.servicetalk.http.api.CharSequences.newAsciiString;
+import static io.servicetalk.buffer.internal.CharSequences.caseInsensitiveHashCode;
+import static io.servicetalk.buffer.internal.CharSequences.contentEqualsIgnoreCase;
+import static io.servicetalk.buffer.internal.CharSequences.equalsIgnoreCaseLower;
+import static io.servicetalk.buffer.internal.CharSequences.newAsciiString;
 import static io.servicetalk.http.api.HeaderUtils.validateCookieNameAndValue;
 import static io.servicetalk.http.api.HeaderUtils.validateCookieTokenAndHeaderName;
 import static io.servicetalk.http.api.HttpSetCookie.SameSite.Lax;
 import static io.servicetalk.http.api.HttpSetCookie.SameSite.None;
 import static io.servicetalk.http.api.HttpSetCookie.SameSite.Strict;
-import static io.servicetalk.utils.internal.CharSequenceUtils.equalsIgnoreCaseLower;
 import static java.lang.Long.parseLong;
 
 /**
