+++ servicetalk-http-api/src/main/java/io/servicetalk/http/api/HttpHeaderValues.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018 Apple Inc. and the ServiceTalk project authors
+ * Copyright © 2018, 2021 Apple Inc. and the ServiceTalk project authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,8 @@
  */
 package io.servicetalk.http.api;
 
-import static io.servicetalk.http.api.CharSequences.newAsciiString;
+
+import static io.servicetalk.buffer.internal.CharSequences.newAsciiString;
 
 /**
  * Common <a href="https://tools.ietf.org/html/rfc7231#section-5">request header values</a> and
